Phase 7 commit 19: Extract LocalBookContentService from MBDC
Lifts the on-disk lifecycle for downloaded book content into a
focused service: deleteLocalContent (×2 overloads),
deleteLocalAudiobookContent, redownloadLCPContentFile (PP-3704), and
the lcpLicenseURL helper.

**Scope:**
- New: `Palace/MyBooks/LocalBookContentService.swift` (~180 LOC w/
  comments). Standard DI-via-init pattern: bookRegistry +
  accountsManager + BookFileManager + FileManager. LCP-specific
  branches (LCPAudiobooks.canOpenBook, LCPPDFs.deletePdfContent, the
  redownload flow) gated #if LCP. Audiobook deletion preserves the
  manifest-decode → toolkit deleteLocalContent sequence for non-LCP
  audiobooks.
- `MyBooksDownloadCenter`: adds `localContentService:
  LocalBookContentService` stored property + nil-default init param
  built from the just-resolved bookRegistry / accountsManager /
  bookFileManager (same instances MBDC owns). Three public methods
  (deleteLocalContent for identifier, deleteLocalContent forBook,
  redownloadLCPContentFile) become 1-line delegators preserving the
  5+ external callers (ReaderService, BookRegistrySync,
  TPPBookRegistryAsync, MyBooksViewModel,
  MyBooksDownloadCenterProtocol). lcpLicenseURL +
  deleteLocalAudiobookContent deleted (now private inside the
  service).
- pbxproj: 6 entries.

**Net effect:**
- MBDC: 2,288 → 2,192 LOC (−96 LOC).
- 128-test smoke green; deletion paths exercised indirectly through
  TPPBookRegistry sync tests + the existing eviction tests.

**Not done:**
- Direct LocalBookContentServiceTests with mocked FileManager —
  defer to follow-up. The fileManager DI seam makes this tractable.
- Mutation testing — palace_mutate.py worktree-path issue continues.

ForgeOS: cs_56c566a9 (init_e7603d4e) — review + testing + release
gates passed; can_release: true.
