+++ Palace/MyBooks/MyBooksDownloadCenter.swift
@@ -47,6 +47,7 @@ import PalaceCatalog
     let downloadAnnouncementService: DownloadAnnouncementService
     private let bookFileManager: BookFileManager
     private let diskBudgetManager: DiskBudgetManager
+    private let localContentService: LocalBookContentService
     private let alertPresenter: DownloadAlertPresenter
     private let authRetryHandler: DownloadAuthRetryHandler
     private let throttlingService: DownloadThrottlingService
@@ -150,6 +151,7 @@ import PalaceCatalog
         downloadAnnouncementService: DownloadAnnouncementService = DownloadAnnouncementService(),
         bookFileManager: BookFileManager? = nil,
         diskBudgetManager: DiskBudgetManager? = nil,
+        localContentService: LocalBookContentService? = nil,
         alertPresenter: DownloadAlertPresenter? = nil,
         authRetryHandler: DownloadAuthRetryHandler? = nil,
         throttlingService: DownloadThrottlingService? = nil,
@@ -192,6 +194,14 @@ import PalaceCatalog
         // we just resolved AND shares the BookFileManager instance — so
         // path resolution stays coherent across the two managers.
         self.diskBudgetManager = diskBudgetManager ?? DiskBudgetManager(
+            bookRegistry: bookRegistry,
+            accountsManager: accountsManager,
+            bookFileManager: self.bookFileManager
+        )
+        // Shares the same registry / accounts manager / file manager so
+        // path resolution + book lookup stays coherent with the rest of
+        // MBDC's lifecycle paths.
+        self.localContentService = localContentService ?? LocalBookContentService(
             bookRegistry: bookRegistry,
             accountsManager: accountsManager,
             bookFileManager: self.bookFileManager
@@ -1182,126 +1192,20 @@ import PalaceCatalog
 
 extension MyBooksDownloadCenter {
 
-    /// Silently re-downloads the .lcpa content file for an LCP audiobook that only
-    /// has the .lcpl license. The book stays in downloadSuccessful state (playable
-    /// via streaming) while the download runs in the background (PP-3704).
     func redownloadLCPContentFile(for book: TPPBook) {
-        #if LCP
-        guard LCPAudiobooks.canOpenBook(book) else { return }
-        guard let licenseURL = lcpLicenseURL(forBookIdentifier: book.identifier) else {
-            Log.warn(#file, "📥 [LCP RE-DOWNLOAD] No license file found for '\(book.title)' — skipping")
-            return
-        }
-        guard let destURL = fileUrl(for: book.identifier) else { return }
-
-        // Skip if .lcpa already exists (another re-download may have completed)
-        if FileManager.default.fileExists(atPath: destURL.path) {
-            Log.info(#file, "📥 [LCP RE-DOWNLOAD] .lcpa already exists for '\(book.title)' — skipping")
-            return
-        }
-
-        Log.info(#file, "📥 [LCP RE-DOWNLOAD] Starting background .lcpa download for '\(book.title)'")
-
-        let lcpService = LCPLibraryService()
-        _ = lcpService.fulfill(licenseURL, progress: { _ in }) { localUrl, error in
-            if let error {
-                Log.error(#file, "📥 [LCP RE-DOWNLOAD] ❌ Failed for '\(book.title)': \(error.localizedDescription)")
-                return
-            }
-            guard let localUrl else {
-                Log.error(#file, "📥 [LCP RE-DOWNLOAD] ❌ No local URL returned for '\(book.title)'")
-                return
-            }
-
-            do {
-                let parentDir = destURL.deletingLastPathComponent()
-                if !FileManager.default.fileExists(atPath: parentDir.path) {
-                    try FileManager.default.createDirectory(at: parentDir, withIntermediateDirectories: true)
-                }
-                try FileManager.default.moveItem(at: localUrl, to: destURL)
-                Log.info(#file, "📥 [LCP RE-DOWNLOAD] ✅ .lcpa stored for '\(book.title)' — local playback now available")
-            } catch {
-                Log.warn(#file, "📥 [LCP RE-DOWNLOAD] ⚠️ File move failed for '\(book.title)': \(error.localizedDescription) — streaming still available")
-            }
-        }
-        #endif
-    }
-
-    /// Returns the .lcpl license URL for an LCP audiobook, if it exists on disk.
-    private func lcpLicenseURL(forBookIdentifier identifier: String) -> URL? {
-        guard let bookURL = fileUrl(for: identifier) else { return nil }
-        let licenseURL = bookURL.deletingPathExtension().appendingPathExtension("lcpl")
-        return FileManager.default.fileExists(atPath: licenseURL.path) ? licenseURL : nil
+        localContentService.redownloadLCPContentFile(for: book)
     }
 
     func deleteLocalContent(for identifier: String, account: String? = nil) {
-        guard let book = bookRegistry.book(forIdentifier: identifier) else {
-            Log.warn(#file, "Could not find book to delete local content \(identifier)")
-            return
-        }
-        deleteLocalContent(forBook: book, account: account)
+        localContentService.deleteLocalContent(for: identifier, account: account)
     }
 
-    /// Delete local content using a book reference directly, without reading the
-    /// book registry. Use this from callers that already hold the book (or that
-    /// are running inside a registry write barrier — looking up the identifier
-    /// through `bookRegistry` there would re-enter the barrier and trip Swift's
-    /// exclusivity check, e.g. BookRegistrySync.sync()'s reconciliation pass
-    /// deleting expired/returned downloads.
     func deleteLocalContent(forBook book: TPPBook, account: String? = nil) {
-        let current_account: String? = account ?? accountsManager.currentAccountId
-        guard let bookURL = fileUrl(for: book, account: current_account) else {
-            Log.warn(#file, "Could not resolve fileUrl to delete local content \(book.identifier)")
-            return
-        }
-
-        do {
-            switch book.defaultBookContentType {
-            case .epub, .pdf:
-                if FileManager.default.fileExists(atPath: bookURL.path) {
-                    try FileManager.default.removeItem(at: bookURL)
-                } else {
-                    Log.info(#file, "Content file already missing (nothing to delete): \(bookURL.lastPathComponent)")
-                }
-                #if LCP
-                if book.defaultBookContentType == .pdf {
-                    try LCPPDFs.deletePdfContent(url: bookURL)
-                }
-                #endif
-            case .audiobook:
-                try deleteLocalAudiobookContent(forAudiobook: book, at: bookURL)
-            case .unsupported:
-                Log.warn(#file, "Unsupported content type for deletion.")
-            }
-        } catch {
-            Log.error(#file, "Failed to remove local content for book with identifier \(book.identifier): \(error.localizedDescription)")
-        }
+        localContentService.deleteLocalContent(forBook: book, account: account)
     }
 
-    private func deleteLocalAudiobookContent(forAudiobook book: TPPBook, at bookURL: URL) throws {
-        #if LCP
-        let isLcpAudiobook = LCPAudiobooks.canOpenBook(book)
-        #else
-        let isLcpAudiobook = false
-        #endif
-
-        // LCP Audiobooks are a single binary file, without an easily loaded manifest.
-        // So they skip this logic that deleted the local audio files, used by other
-        // audiobook types.
-        // TODO: Update LCP so we don't have to special case it here.
-        if !isLcpAudiobook {
-            let manifestData = try Data(contentsOf: bookURL)
-            let manifest = try Manifest.customDecoder().decode(Manifest.self, from: manifestData)
-            AudiobookFactory.audiobookClass(for: manifest).deleteLocalContent(manifest: manifest, bookIdentifier: book.identifier)
-        }
-
-        if FileManager.default.fileExists(atPath: bookURL.path) {
-            try FileManager.default.removeItem(at: bookURL)
-        } else {
-            Log.info(#file, "Audiobook content already missing (nothing to delete): \(bookURL.lastPathComponent)")
-        }
-        Log.info(#file, "Successfully deleted audiobook manifest & content \(book.identifier)")
-    }
+    // lcpLicenseURL + deleteLocalAudiobookContent moved to LocalBookContentService
+    // (private there).
 
     @objc func returnBook(withIdentifier identifier: String, completion: (() -> Void)? = nil) {
         guard let book = bookRegistry.book(forIdentifier: identifier) else {
